Finished to implement the coordinates and the new mechanics
Redone everything that was failing like the sliding not working or the colliders bad setted up. Put on example coordinates for Alexis to load his art and finished redoing the example code for the player the background and the animations. Simply ignore the examples there Alexis is going to change to his own